Fix bug - Now cast power operations to float instead of doubles.

+++ Compiler/src/Number.cpp
@@ -141,19 +141,19 @@ std::pair<Number*, Error*> Number::power(Number* other)
 	result->context = context;
 
 	if (value.index() == 0 && other->value.index() == 0) {
-		try { result->value = pow(std::get<float>(value), std::get<float>(other->value)); }
+		try { result->value = (float)pow(std::get<float>(value), std::get<float>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 1) {
-		try { result->value = pow((float)std::get<int>(value), std::get<int>(other->value)); }
+		try { result->value = (float)pow((float)std::get<int>(value), std::get<int>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 0 && other->value.index() == 1) {
-		try { result->value = pow(std::get<float>(value), (float)std::get<int>(other->value)); }
+		try { result->value = (float)pow(std::get<float>(value), (float)std::get<int>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
-		try { result->value = pow((float)std::get<int>(value), std::get<float>(other->value)); }
+		try { result->value = (float)pow((float)std::get<int>(value), std::get<float>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
